PgSQL

+++ abdata/abdata/Database.h
@@ -54,7 +54,7 @@ public:
     virtual void FreeResult(DBResult* res);
     virtual uint64_t GetLastInsertId() = 0;
     virtual std::string EscapeString(const std::string& s) = 0;
-    virtual std::string EscapeBlob(const char* s, uint32_t length) = 0;
+    virtual std::string EscapeBlob(const char* s, size_t length) = 0;
 };
 
 class DBResult : public std::enable_shared_from_this<DBResult>
